+++ packages/website/ts/components/ui/filled_image.tsx
